post formuyla ürün silme

+++ FormsApp/Controllers/HomeController.cs
@@ -140,6 +140,23 @@ public class HomeController : Controller
             return NotFound();
         }
 
+        return View("DeleteConfirm", entity);
+    }
+
+    [HttpPost]
+    public IActionResult Delete(int id, int ProductId)
+    {
+        if(id != ProductId)
+        {
+            return NotFound();
+        }
+
+        var entity = Repository.Products.FirstOrDefault(p => p.ProductId == ProductId);
+        if(entity == null)
+        {
+            return NotFound();
+        }
+
         Repository.DeleteProduct(entity);
         return RedirectToAction("Index");
     }
